Add argpase and move stuff around
Add argparse, move and rename main script, add pixi.
