+++ .envrc
@@ -1,0 +1,3 @@
+watch_file pixi.lock 
+eval "$(pixi shell-hook)" 
+
